feat : a test that ensures when the loadFeed         function is called and completes its work, there is at least         a single element within the feed's container.
Here before we check if there is at least a single entry
after we call the loadFeed() function,
we should to have some way told us if loadFeed() has completed,
so we pass done() to each function before they invoke,
so that when the function loadFeed() completed done() will returned.
Before the loadFeed() is invoked there is no element
has the class 'entry' in it,
so we need to see if there is an element with that class
after the loadFeed() is invoked to make sure there is an entry
in the feed's container.