feat : a test that ensures when a new feed is loaded         by the loadFeed function that the content actually changes.
Here we have tow variables firstEntries and secondEntries to hold
the value of the entries after first loadFeed
and the value of the entries after the second loadFeed.

we sould call loadFeed twice and make sure it has been completed
and assign the first call result to firstEntries
and the second call result to secondEntries.

Lastly we need to decide if the content of the feeds are changed,
so we need to compare between firstEntries and secondEntries and
they should not be the same.